+++ envoy/datadog_checks/envoy/parser.py
@@ -1,37 +1,71 @@
-from .errors import UnknownMetric
+from six.moves import range, zip
+
+from .errors import UnknownMetric, UnknownTags
 from .metrics import METRIC_PREFIX, METRIC_TREE, METRICS
 
 
-def parse_metric(metric):
+def parse_metric(metric, metric_mapping=METRIC_TREE):
     """Takes a metric formatted by Envoy and splits it into a unique
     metric name. Returns the unique metric name, a list of tags, and
     the name of the submission method.
 
     Example:
         'listener.0.0.0.0_80.downstream_cx_total' ->
-        ('listener.downstream_cx_total', ['address:0.0.0.0:80'], 'count')
+        ('listener.downstream_cx_total', ['address:0.0.0.0_80'], 'count')
     """
     metric_parts = []
+    tag_names = []
     tag_values = []
-    mapping = METRIC_TREE
+    tag_builder = []
+    unknown_tags = []
+    num_tags = 0
+    minimum_tag_length = 0
 
     for metric_part in metric.split('.'):
-        if metric_part in mapping:
+        if metric_part in metric_mapping and num_tags >= minimum_tag_length:
+            # Rebuild any built up tags whenever we encounter a known metric part.
+            if tag_builder:
+                for tags in metric_mapping['|_tags_|']:
+                    if num_tags >= len(tags):
+                        break
+
+                constructed_tags = construct_tags(tag_builder, len(tags))
+
+                if tags:
+                    tag_names.extend(tags)
+                    tag_values.extend(constructed_tags)
+                else:
+                    unknown_tags.extend(constructed_tags)
+
+                num_tags = 0
+
             metric_parts.append(metric_part)
-            mapping = mapping[metric_part]
+            metric_mapping = metric_mapping[metric_part]
+            minimum_tag_length = len(metric_mapping['|_tags_|'][-1])
         else:
-            tag_values.append(metric_part)
+            tag_builder.append(metric_part)
+            num_tags += 1
 
     metric = '.'.join(metric_parts)
     if metric not in METRICS:
         raise UnknownMetric
 
-    tag_names = METRICS[metric]['tags']
-    if len(tag_values) != len(tag_names):
-        tag_values = reassemble_addresses(tag_values)
+    # Rebuild any trailing tags
+    if tag_builder:
+        for tags in metric_mapping['|_tags_|']:
+            if num_tags >= len(tags):
+                break
 
-        if len(tag_values) != len(tag_names):
-            raise UnknownMetric
+        constructed_tags = construct_tags(tag_builder, len(tags))
+
+        if tags:
+            tag_names.extend(tags)
+            tag_values.extend(constructed_tags)
+        else:
+            unknown_tags.extend(constructed_tags)
+
+    if unknown_tags:
+        raise UnknownTags('{}'.format('|||'.join(unknown_tags)))
 
     tags = [
         '{}:{}'.format(tag_name, tag_value)
@@ -41,33 +75,16 @@ def parse_metric(metric):
     return METRIC_PREFIX + metric, tags, METRICS[metric]['method']
 
 
-def reassemble_addresses(seq):
-    """Takes a sequence of strings and combines any sub-sequence that looks
-    like an IPv4 address into a single string.
+def construct_tags(tag_builder, num_tags):
+    # First fill in all trailing slots with one tag.
+    tags = [tag_builder.pop() for _ in range(num_tags - 1)]
 
-    Example:
-        ['listener', '0', '0', '0', '0_80', downstream_cx_total'] ->
-        ['listener', '0.0.0.0:80', 'downstream_cx_total']
-    """
-    reassembled = []
-    prev = ''
+    # Merge any excess tag parts.
+    if tag_builder:
+        tags.append('.'.join(tag_builder))
 
-    for s in seq:
-        if prev.isdigit():
-            try:
-                end, port = s.split('_')
-            except ValueError:
-                end, port = '', ''
+    # Once the builder has been used, clear its contents.
+    del tag_builder[:]
 
-            if s.isdigit():
-                reassembled[-1] += '.{}'.format(s)
-            elif end.isdigit() and port.isdigit():
-                reassembled[-1] += '.{}:{}'.format(end, port)
-            else:
-                reassembled.append(s)
-        else:
-            reassembled.append(s)
-
-        prev = s
-
-    return reassembled
+    # Return an iterator in the original order.
+    return reversed(tags)