+++ exchange_server/datadog_checks/exchange_server/__init__.py
@@ -1,7 +1,10 @@
-from . import exchange_server
+# (C) Datadog, Inc. 2018
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .__about__ import __version__
+from .exchange_server import ExchangeCheck
 
-ExchangeCheck = exchange_server.ExchangeCheck
-
-__version__ = "1.0.0"
-
-__all__ = ['exchange_server']
+__all__ = [
+    '__version__',
+    'ExchangeCheck'
+]