+++ postgres/tests/common.py
@@ -1,0 +1,10 @@
+# (C) Datadog, Inc. 2010-2018
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
+import os
+
+HOST = os.getenv('DOCKER_HOSTNAME', 'localhost')
+PORT = '5432'
+USER = 'datadog'
+PASSWORD = 'datadog'
+DB_NAME = 'datadog_test'