+++ snmp/datadog_checks/snmp/__init__.py
@@ -2,6 +2,6 @@ from . import snmp
 
 SnmpCheck = snmp.SnmpCheck
 
-__version__ = "1.3.1"
+__version__ = "1.4.0"
 
 __all__ = ['snmp']