+++ docker_daemon/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG - docker_daemon
 
+1.10.0 / Unreleased
+==================
+### Changes
+
+* [IMPROVEMENT] Add docker.cpu.shares metric. See [#1358][]
+
 1.9.0 / 2018-03-23
 ==================
 ### Changes