+++ envoy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG - Envoy
 
+1.1.0 / Unreleased
+==================
+
+### Changes
+
+* [FEATURE] add newly-documented metrics. See #1326
+* [IMPROVEMENT] tags can now contain the dot metric delimiter itself. See #1404
+
 1.0.0 / 2018-03-23
 ==================