+++ oracle/CHANGELOG.md
@@ -1,7 +1,21 @@
 # CHANGELOG - oracle
 
-1.0.1 / 2018-03-23 
+1.2.0 / Unreleased
 ==================
+
+### Changes
+
+* [FEATURE] adds metric `oracle.tablespace.offline`. See #1402
+* [BUGFIX] fix for DB with offline tablespace. See #1402
+
+1.1.0 / Unreleased
+==================
+
+### Changes
+
+* [FEATURE] adds custom tag support to service checks.
+
+1.0.1 / Unreleased 
 
 ### Notes